refactoring the create and delete users controllers
